ext_proc::envoy::config::core

Module v3

Modules

address

Nested message and enum types in Address.

async_data_source

Nested message and enum types in AsyncDataSource.

data_source

Nested message and enum types in DataSource.

envoy_internal_address

Nested message and enum types in EnvoyInternalAddress.

header_value_option

Nested message and enum types in HeaderValueOption.

http_uri

Nested message and enum types in HttpUri.

key_value_append

Nested message and enum types in KeyValueAppend.

node

Nested message and enum types in Node.

retry_policy

Nested message and enum types in RetryPolicy.

socket_address

Nested message and enum types in SocketAddress.

socket_option

Nested message and enum types in SocketOption.

transport_socket

Nested message and enum types in TransportSocket.

Structs

Address

Addresses specify either a logical or physical address and port, which are used to tell Envoy where to bind/listen, connect to upstream and find management servers.

AsyncDataSource

Async data source which support async data fetch.

BackoffStrategy

Configuration defining a jittered exponential back off strategy.

BindConfig

[#next-free-field: 7]

BuildVersion

BuildVersion combines SemVer version of extension with free-form build information (i.e. ‘alpha’, ‘private-build’) as a set of strings.

CidrRange

CidrRange specifies an IP Address and a prefix length to construct the subnet mask for a CIDR <<https://tools.ietf.org/html/rfc4632>_> range.

ControlPlane

Identifies a specific ControlPlane instance that Envoy is connected to.

DataSource

Data source consisting of a file, an inline value, or an environment variable. [#next-free-field: 6]

EnvoyInternalAddress

The address represents an envoy internal listener. [#comment: TODO(asraa): When address available, remove workaround from test/server/server_fuzz_test.cc:30.]

Extension

Version and identification for an Envoy extension. [#next-free-field: 7]

ExtraSourceAddress

HeaderMap

Wrapper for a set of headers.

HeaderValue

Header name/value pair.

HeaderValueOption

Header name/value pair plus option to control append behavior.

HttpUri

Envoy external URI descriptor

KeyValue

Please use :ref:KeyValuePair <envoy_api_msg_config.core.v3.KeyValuePair> instead. [#not-implemented-hide:]

KeyValueAppend

Key/value pair plus option to control append behavior. This is used to specify key/value pairs that should be appended to a set of existing key/value pairs.

KeyValueMutation

Key/value pair to append or remove.

KeyValuePair

Locality

Identifies location of where either Envoy runs or where upstream hosts run.

Metadata

Metadata provides additional inputs to filters based on matched listeners, filter chains, routes and endpoints. It is structured as a map, usually from filter name (in reverse DNS format) to metadata specific to the filter. Metadata key-values for a filter are merged as connection and request handling occurs, with later values for the same key overriding earlier values.

Node

Identifies a specific Envoy instance. The node identifier is presented to the management server, which may use this identifier to distinguish per Envoy configuration for serving. [#next-free-field: 13]

Pipe

QueryParameter

Query parameter name/value pair.

RemoteDataSource

The message specifies how to fetch data from remote and how to verify it.

RetryPolicy

The message specifies the retry policy of remote data source when fetching fails. [#next-free-field: 7]

RuntimeDouble

Runtime derived double with a default when not specified.

RuntimeFeatureFlag

Runtime derived bool with a default when not specified.

RuntimeFractionalPercent

Runtime derived FractionalPercent with defaults for when the numerator or denominator is not specified via a runtime key.

RuntimePercent

Runtime derived percentage with a default when not specified.

RuntimeUInt32

Runtime derived uint32 with a default when not specified.

SocketAddress

[#next-free-field: 7]

SocketOption

Generic socket option message. This would be used to set socket options that might not exist in upstream kernels or precompiled Envoy binaries.

SocketOptionsOverride

TcpKeepalive

TransportSocket

Configuration for transport socket in :ref:listeners <config_listeners> and :ref:clusters <envoy_v3_api_msg_config.cluster.v3.Cluster>. If the configuration is empty, a default transport socket implementation and configuration will be chosen based on the platform and existence of tls_context.

TypedExtensionConfig

Message type for extension configuration. [#next-major-version: revisit all existing typed_config that doesn’t use this wrapper.].

WatchedDirectory

A directory that is watched for changes, e.g. by inotify on Linux. Move/rename events inside this directory trigger the watch.

Enums

RequestMethod

HTTP request method.

RoutingPriority

Envoy supports :ref:upstream priority routing <arch_overview_http_routing_priority> both at the route and the virtual cluster level. The current priority implementation uses different connection pool and circuit breaking settings for each priority level. This means that even for HTTP/2 requests, two physical connections will be used to an upstream host. In the future Envoy will likely support true HTTP/2 priority over a single upstream connection.

TrafficDirection

Identifies the direction of the traffic relative to the local Envoy.