ext_proc::envoy::service::ext_proc

Module v3

Modules

body_mutation

Nested message and enum types in BodyMutation.

common_response

Nested message and enum types in CommonResponse.

external_processor_client

Generated client implementations.

external_processor_server

Generated server implementations.

processing_request

Nested message and enum types in ProcessingRequest.

processing_response

Nested message and enum types in ProcessingResponse.

Structs

BodyMutation

This message specifies the body mutation the server sends to Envoy.

BodyResponse

This message is sent by the external server to Envoy after HttpBody was sent to it.

CommonResponse

This message contains common fields between header and body responses. [#next-free-field: 6]

GrpcStatus

This message specifies a gRPC status for an ImmediateResponse message.

HeaderMutation

Change HTTP headers or trailers by appending, replacing, or removing headers.

HeadersResponse

This message is sent by the external server to Envoy after HttpHeaders was sent to it.

HttpBody

This message is sent to the external server when the HTTP request and response bodies are received.

HttpHeaders

This message is sent to the external server when the HTTP request and responses are first received.

HttpTrailers

This message is sent to the external server when the HTTP request and response trailers are received.

ImmediateResponse

This message causes the filter to attempt to create a locally generated response, send it downstream, stop processing additional filters, and ignore any additional messages received from the remote server for this request or response. If a response has already started, then this will either ship the reply directly to the downstream codec, or reset the stream. [#next-free-field: 6]

ProcessingRequest

This represents the different types of messages that Envoy can send to an external processing server. [#next-free-field: 11]

ProcessingResponse

For every ProcessingRequest received by the server with the observability_mode field set to false, the server must send back exactly one ProcessingResponse message. [#next-free-field: 11]

StreamedBodyResponse

[#not-implemented-hide:] The body response message corresponding to FULL_DUPLEX_STREAMED body mode.

TrailersResponse

This message is sent by the external server to Envoy after HttpTrailers was sent to it.