ext_proc::envoy::type

Module v3

Modules

fractional_percent

Nested message and enum types in FractionalPercent.

Structs

FractionalPercent

A fractional percentage is used in cases in which for performance reasons performing floating point to integer conversions during randomness calculations is undesirable. The message includes both a numerator and denominator that together determine the final fractional value.

HttpStatus

HTTP status.

Percent

Identifies a percentage, in the range [0.0, 100.0].

SemanticVersion

Envoy uses SemVer (https://semver.org/). Major/minor versions indicate expected behaviors and APIs, the patch version field is used only for security fixes and can be generally ignored.

Enums

StatusCode

HTTP response codes supported in Envoy. For more details: https://www.iana.org/assignments/http-status-codes/http-status-codes.xhtml